+++ klum-ast/src/main/java/com/blackbuild/klum/ast/util/layer3/DefaultValuesCheck.java
@@ -1,0 +1,50 @@
+/*
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2015-2025 Stephan Pauxberger
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+package com.blackbuild.klum.ast.util.layer3;
+
+import com.blackbuild.klum.cast.checks.impl.KlumCastCheck;
+import com.blackbuild.klum.cast.checks.impl.ValidationException;
+import org.codehaus.groovy.ast.AnnotatedNode;
+import org.codehaus.groovy.ast.AnnotationNode;
+import org.codehaus.groovy.ast.ClassNode;
+
+import java.lang.annotation.Annotation;
+
+public class DefaultValuesCheck extends KlumCastCheck<Annotation> {
+    @Override
+    protected void doCheck(AnnotationNode annotationToCheck, AnnotatedNode target) throws ValidationException {
+        boolean controlAnnotationHasValuesMapping = annotationToCheck.getMembers().containsKey("valueTarget");
+        ClassNode targetAnnotation = (ClassNode) target;
+        boolean targetAnnotationHasValueMember = !targetAnnotation.getMethods("value").isEmpty();
+
+        if (controlAnnotationHasValuesMapping && !targetAnnotationHasValueMember)
+            throw new ValidationException(
+                String.format("DefaultValues has a 'valueTarget' member, but the target annotation %s does not have a 'value' member", targetAnnotation.getName()), target);
+
+        if (!controlAnnotationHasValuesMapping && targetAnnotationHasValueMember)
+            throw new ValidationException(
+                String.format("Target annotation %s does have a 'value' member, but DefaultValues does not have a 'valueTarget' member", targetAnnotation.getName()), target);
+
+    }
+}