+++ CHANGES.md
@@ -35,7 +35,7 @@
 - DefaultValues annotations provide an annotation based way to set default values in Layer3 scenarios (see [Default Values](https://github.com/klum-dsl/klum-ast/wiki/Default-Values#DefaultValues-annotation) and [#361](https://github.com/klum-dsl/klum-ast/issues/361))
 - `@Cluster`-Fields create now a factory closure for that field, containing only the cluster members (see [#365](https://github.com/klum-dsl/klum-ast/issues/365))
 - `applyLater` methods for all objects that can be used to hook closures to be applied in later phases (see [Model Phases](https://github.com/klum-dsl/klum-ast/wiki/Model-Phases))
-- Validations have additional levels (WARNING, DEPRECATION and INFO) that can be set for each individual validation. Non error validations to lead to failure in the model (see [#145](https://github.com/klum-dsl/klum-ast/issues/365) and [Validation](https://github.com/klum-dsl/klum-ast/wiki/Validation))
+- Validations have additional levels (WARNING, DEPRECATION and INFO) that can be set for each individual validation. Non error validations to lead to failure in the model (see [#145](https://github.com/klum-dsl/klum-ast/issues/145) and [Validation](https://github.com/klum-dsl/klum-ast/wiki/Validation))
 
 ## Improvements
 - Creator classes also support methods creating multiple instances at once (see [#319](https://github.com/klum-dsl/klum-ast/issues/319))
@@ -52,6 +52,7 @@
 - `@Cluster` can also be placed on fields, which will be converted into a getter method (see [#366](https://github.com/klum-dsl/klum-ast/issues/366))
 - `@Cluster` can be combined with `@AutoCreate` to auto create all members of the annotated cluster (see [#363](https://github.com/klum-dsl/klum-ast/issues/363))
 - Templates that where active during an object's creation are now stored in the proxy of that object and will be applied to any object created by that object in later phases (usually AutoCreate) (see [#377](https://github.com/klum-dsl/klum-ast/issues/377))
+- `@DefaultValues` has a `valueTarget` member that is used to remap the `value` member of the targeted annotation to a specific field (see [#370](https://github.com/klum-dsl/klum-ast/issues/370))
 
 ## Deprecations (see [Migration](https://github.com/klum-dsl/klum-ast/wiki/Migration)):
   - The `@Validation` annotation is deprecated. Use `@Validate` on class level instead.